requirements.txt
# Core dependencies
alpaca-py>=0.10.0
polars>=0.20.0
python-dotenv>=1.0.0
numpy>=1.24.0
pytz>=2023.3
pyarrow>=14.0.0

# ML/PyTorch
torch>=2.0.0
scikit-learn>=1.3.0

# Visualization (optional, for scripts)
matplotlib>=3.7.0
